+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# ChangeLog
+
 ## 1.0.0
 
 - Initial version.
@@ -7,3 +9,9 @@
 - Correct `README.md` examples
 - Correct Repository Url in `pubspec.yaml`
 
+## 1.1.0
+
+- Add filtered listeners
+- Add typed listeners
+- Rename `Event.convertTo` to `Event.linkTo`
+- Create new class `EventComponent`